+++ pysat/instruments/icon_fuv.py
@@ -3,13 +3,13 @@
 CONnection Explorer (ICON) satellite.  Accesses local data in
 netCDF format.
 
-Parameters
+Properties
 ----------
-platform : string
+platform
     'icon'
-name : string
+name
     'fuv'
-tag : string
+tag
     None supported
 
 Warnings
@@ -29,6 +29,7 @@ Example
 By default, pysat removes the ICON level tags from variable names, ie,
 ICON_L27_Ion_Density becomes Ion_Density.  To retain the original names, use
 ::
+
     fuv = pysat.Instrument(platform='icon', name='fuv', tag=day',
                            keep_original_names=True)
 
@@ -99,7 +100,7 @@ def init(self):
 
     Parameters
     -----------
-    inst : (pysat.Instrument)
+    inst : pysat.Instrument
         Instrument class object
 
     """
@@ -118,9 +119,8 @@ def default(inst):
 
     Parameters
     -----------
-    inst : (pysat.Instrument)
+    inst : pysat.Instrument
         Instrument class object
-
 
     """
 
@@ -172,6 +172,7 @@ def load(fnames, tag=None, sat_id=None, keep_original_names=False):
     Examples
     --------
     ::
+
         inst = pysat.Instrument('icon', 'fuv')
         inst.load(2020, 1)
 
@@ -194,19 +195,11 @@ def load(fnames, tag=None, sat_id=None, keep_original_names=False):
 def clean(inst):
     """Provides data cleaning based upon clean_level.
 
-    clean_level is set upon Instrument instantiation to
-    one of the following:
-
-    'Clean'
-    'Dusty'
-    'Dirty'
-    'None'
-
     Routine is called by pysat, and not by the end user directly.
 
     Parameters
     -----------
-    inst : (pysat.Instrument)
+    inst : pysat.Instrument
         Instrument class object, whose attribute clean_level is used to return
         the desired level of data selectivity.