+++ pysat/instruments/methods/general.py
@@ -107,9 +107,9 @@ def convert_timestamp_to_datetime(inst, sec_mult=1.0, epoch_name='Epoch'):
 
     """
 
-    inst.data[epoch_name] = \
-        pds.to_datetime([dt.datetime.utcfromtimestamp(x * sec_mult)
-                         for x in inst.data[epoch_name]])
+    inst.data[epoch_name] = pds.to_datetime(
+        [dt.datetime.utcfromtimestamp(x * sec_mult)
+         for x in inst.data[epoch_name]])
     return